refactor: :truck: nativeBridge, nativeMessageReceiver util로 이동
