+++ src/shared/utils/nativeMessageReceiver.ts
